+++ src/Engine/Engine.h
@@ -249,17 +249,6 @@ public:
      */
     void        deleteAllInstrumentations();
 
-    /*! Obtain the analysis of an instruction metadata. Analysis results are cached in the VM.
-     *  The validity of the returned pointer is only guaranteed until the end of the callback, else
-     *  a deepcopy of the structure is required.
-     *
-     * @param[in] instMetadata Metadata to analyze.
-     * @param[in] type         Properties to retrieve during analysis.
-     *
-     * @return A InstAnalysis structure containing the analysis result.
-     */
-    const InstAnalysis* analyzeInstMetadata(const InstMetadata* instMetadata, AnalysisType type);
-
     /*! Expose current ExecBlock
      *
      * @return A pointer to current ExecBlock
@@ -279,6 +268,17 @@ public:
      * @return True if basic block has been inserted in cache.
      */
     bool precacheBasicBlock(rword pc);
+
+    /*! Return an InstAnalysis for a cached instruction.
+     * The pointer may be invalid by any noconst method call.
+     *
+     * @param[in] address Start address of the instruction
+     * @param[in] type    type of the Analysis
+     *
+     * @return A pointer to the Analysis or a null pointer if the instruction isn't
+     *    in the cache.
+     */
+    const InstAnalysis* getInstAnalysis(rword address, AnalysisType type) const;
 
     /*! Clear a specific address range from the translation cache.
      *