+++ src/ExecBlock/ExecBlock.cpp
@@ -26,6 +26,7 @@
 #include "Patch/PatchRules_Target.h"
 #include "Patch/RelocatableInst.h"
 #include "Utility/Assembly.h"
+#include "Utility/InstAnalysis_prive.h"
 #include "Utility/LogSys.h"
 #include "Utility/System.h"
 
@@ -55,7 +56,7 @@ RelocatableInst::SharedPtrVec ExecBlock::execBlockPrologue = RelocatableInst::Sh
 RelocatableInst::SharedPtrVec ExecBlock::execBlockEpilogue = RelocatableInst::SharedPtrVec();
 void (*ExecBlock::runCodeBlockFct)(void*) = NULL;
 
-ExecBlock::ExecBlock(Assembly &assembly, VMInstanceRef vminstance) : vminstance(vminstance), assembly(assembly) {
+ExecBlock::ExecBlock(const Assembly &assembly, VMInstanceRef vminstance) : vminstance(vminstance), assembly(assembly) {
     // Allocate memory blocks
     std::error_code ec;
     // iOS now use 16k superpages, but as JIT mecanisms are totally differents
@@ -286,7 +287,9 @@ SeqWriteResult ExecBlock::writeSequence(std::vector<Patch>::const_iterator seqIt
         }
         else {
             // Complete instruction was written, we add the metadata
-            instMetadata.push_back(seqIt->metadata);
+            // Move the analysis of the instruction in the cached metadata
+            instMetadata.push_back(seqIt->metadata.lightCopy());
+            instMetadata.back().analysis.reset(seqIt->metadata.analysis.release());
             // Register instruction
             instRegistry.push_back(InstInfo {seqID, static_cast<uint16_t>(rollbackOffset)});
             // Update indexes
@@ -402,6 +405,11 @@ rword ExecBlock::getInstAddress(uint16_t instID) const {
 const llvm::MCInst* ExecBlock::getOriginalMCInst(uint16_t instID) const {
     Require("ExecBlock::getOriginalMCInst", instID < instMetadata.size());
     return &instMetadata[instID].inst;
+}
+
+const InstAnalysis* ExecBlock::getInstAnalysis(uint16_t instID, AnalysisType type) const {
+    Require("ExecBlock::getInstAnalysis", instID < instMetadata.size());
+    return analyzeInstMetadata(instMetadata[instID], type, assembly);
 }
 
 uint16_t ExecBlock::getSeqID(rword address) const {